fix: better matching for goodreads vs libby titles
Prior to this, I was doing an exact match on goodreads vs
libby titles when searching. This was shortsighted, since
many goodreads titles are longer than what shows up in Libby.

e.g. In GR we might have
"Atlas of the Heart: Mapping Meaningful Connection and the Language of Human Experience"

but in Libby this just shows as
"Atlas of the Heart"

Relatedly, sometimes including all the text from the GR title
in my overdrive search would lead to no results at all.

e.g. for "Dune (Dune, #1)", we only get a single result for a
spanish version of Dune. If we simply search for "Dune" instead,
we get expected results.

In this commit I also drive-by cleaned up some logging.